+++ test/fixtures/app/index.js
@@ -5,5 +5,6 @@ import login from './login'
 const html = fn()
 
 assert(html.contains('Article'))
+assert(html.contains('ThisIsThePugPlugin'))
 
 login()