+++ test/run.js
@@ -1,10 +1,8 @@
+const rollup = require('rollup').rollup
+const assert = require('assert')
+const _pug   = require('../')
 
-var rollup = require('rollup').rollup
-var buble  = require('rollup-plugin-buble')
-var assert = require('assert')
-var _pug   = require('../')
-
-var WRITE_BUNDLE = true
+const WRITE_BUNDLE = true
 
 process.chdir(__dirname)
 
@@ -18,25 +16,25 @@ try {
 
 function writeBundle (bundle, name) {
   if (WRITE_BUNDLE && name) {
-    bundle.write({ format: 'cjs', dest: '~tmp/' + name + '.js' })
+    bundle.write({ format: 'cjs', file: `~tmp/${name}.js` })
   }
 }
 
 
 function executeBundle (bundle, name) {
-  var result = bundle.generate({
-    format: 'cjs'
+  bundle.generate({
+    format: 'cjs',
+  }).then((result) => {
+    if (name) {
+      writeBundle(bundle, name)
+    }
+
+    // eslint-disable-next-line no-new-func
+    const fn = new Function('require', 'module', 'assert', result.code)
+    const module = {}
+
+    fn(require, module, assert)
   })
-
-  if (name) {
-    writeBundle(bundle, name)
-  }
-
-  // eslint-disable-next-line no-new-func
-  var fn = new Function('require', 'module', 'assert', result.code)
-  var module = {}
-
-  fn(require, module, assert)
 }
 
 
@@ -44,220 +42,202 @@ describe('rollup-plugin-pug', function () {
 
   it('compiles pug templates to funcions', function (done) {
     rollup({
-      entry: 'fixtures/basic/main.js',
-      plugins: [_pug(), buble()]
+      input: 'fixtures/basic/main.js',
+      plugins: [_pug()],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'basic')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'basic'))
+      .then(done, done)
   })
 
   it('inserts the includes into the template', function (done) {
     rollup({
-      entry: 'fixtures/include/main.js',
-      plugins: [_pug(), buble()]
+      input: 'fixtures/include/main.js',
+      plugins: [_pug()],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'include')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'include'))
+      .then(done, done)
   })
 
   it('can use pug options', function (done) {
     rollup({
-      entry: 'fixtures/options/main.js',
+      input: 'fixtures/options/main.js',
       plugins: [_pug({
         doctype: 'strict',
-        pretty: '\t'
-      })]
+        pretty: '\t',
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'options')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'options'))
+      .then(done, done)
   })
 
   it('allows custom globals in addition to the default ones', function (done) {
     rollup({
-      entry: 'fixtures/custom_globals/main.js',
+      input: 'fixtures/custom_globals/main.js',
       plugins: [_pug({
         sourceMap: false,
-        globals: ['require', 'String'] // String is default, coverage only
-      })]
+        globals: ['require', 'String'], // String is default, coverage only
+      })],
     })
-    .then(function (bundle) {
-      var result = bundle.generate({
-        format: 'es'
+      .then((bundle) => bundle.generate({ format: 'es' }))
+      .then((result) => {
+        //writeBundle(bundle, 'custom_globals')
+        assert(result.code.indexOf('typeof require2!') > -1)
+        assert(result.code.indexOf('typeof require!') < 0)
+        assert(result.code.indexOf('typeof Math!') < 0)
       })
-      writeBundle(bundle, 'custom_globals')
-
-      assert(result.code.indexOf('typeof require2!') > -1)
-      assert(result.code.indexOf('typeof require!') < 0)
-      assert(result.code.indexOf('typeof Math!') < 0)
-
-      done()
-    })
-    .catch(done)
+      .then(done, done)
   })
 
   it('can use .pug extensions', function (done) {
     rollup({
-      entry: 'fixtures/basic/main.js',
+      input: 'fixtures/basic/main.js',
       plugins: [_pug({
-        extensions: 'pug'
-      })]
+        extensions: 'pug',
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle)
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle))
+      .then(done, done)
   })
 
   it('can use the `include: []` option with `extensions: "*"`', function (done) {
     rollup({
-      entry: 'fixtures/basic/main.js',
+      input: 'fixtures/basic/main.js',
       plugins: [_pug({
         include: '**/*.pug',
-        extensions: '*'
-      })]
+        extensions: '*',
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle)
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle))
+      .then(done, done)
   })
 
   it('move imports after a dash out of the pug function', function (done) {
     rollup({
-      entry: 'fixtures/imports/main.js',
-      plugins: [_pug()]
+      input: 'fixtures/imports/main.js',
+      plugins: [_pug()],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'imports')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'imports'))
+      .then(done, done)
   })
 
   it('has static compilation', function (done) {
     rollup({
-      entry: 'fixtures/precompile/main.js',
+      input: 'fixtures/precompile/main.js',
       plugins: [_pug({
         locals: { name: 'pug', other: 'other' },
         option_local: 'option_local',
-      })]
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'precompile')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'precompile'))
+      .then(done, done)
   })
 
   it('static compilation ignores compileDebug', function (done) {
     rollup({
-      entry: 'fixtures/precompile/main.js',
+      input: 'fixtures/precompile/main.js',
       plugins: [_pug({
         locals: { name: 'pug', other: 'other' },
         option_local: 'option_local',
-        compileDebug: true
-      })]
+        compileDebug: true,
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'precompile_debug')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'precompile_debug'))
+      .then(done, done)
   })
 
-  it('defaults to the correct path based on rollup entry', function (done) {
+  it('static compilation watch includes', function (done) {
     rollup({
-      entry: '../test/fixtures/app/index.js',
-      plugins: [_pug()]
+      input: 'fixtures/precompile_inc/main.js',
+      plugins: [_pug({
+        locals: { name: 'pug', other: 'other' },
+        option_local: 'option_local',
+        compileDebug: true,
+      })],
     })
-    .then(function (bundle) {
-      var result = bundle.generate({
-        format: 'es'
+      .then((bundle) => executeBundle(bundle, 'precompile_inc'))
+      .then(done, done)
+  })
+
+  it('defaults to the correct path based on rollup input', function (done) {
+    rollup({
+      input: '../test/fixtures/app/index.js',
+      plugins: [_pug()],
+    })
+      .then((bundle) => {
+        return bundle.generate({ format: 'es' }).then((result) => {
+          assert(~result.code.indexOf('Article'))
+        })
       })
-      writeBundle(bundle, 'app')
+      .then(done, done)
+  })
 
-      assert(~result.code.indexOf('Article'))
-
-      done()
+  it('gets the base path through the `basedir` option', function (done) {
+    rollup({
+      input: '../test/fixtures/app/index.js',
+      plugins: [_pug({
+        basedir: 'fixtures/app',
+      })],
     })
-    .catch(done)
+      .then((bundle) => {
+        return bundle.generate({ format: 'es' }).then((result) => {
+          writeBundle(bundle, 'app')
+          assert(~result.code.indexOf('Article'))
+        })
+      })
+      .then(done, done)
   })
 
   it('does not import the Pug runtime if `pugRuntime` is falsy', function (done) {
     rollup({
-      entry: 'fixtures/no_runtime/main.js',
+      input: 'fixtures/no_runtime/main.js',
       plugins: [_pug({
-        pugRuntime: false
-      })]
+        pugRuntime: false,
+      })],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'no_runtime')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'no_runtime'))
+      .then(done, done)
   })
 
   it('allows import a custom Pug runtime with `pugRuntime: "[custom-runtime]"`', function (done) {
     rollup({
-      entry: 'fixtures/custom_runtime/main.js',
+      input: 'fixtures/custom_runtime/main.js',
       plugins: [_pug({
         pugRuntime: 'pug-runtime',
       })],
-      external: ['pug-runtime']
+      external: ['pug-runtime'],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'custom_runtime')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'custom_runtime'))
+      .then(done, done)
   })
 
   it('can import custom runtime w/ `import pug from...` and `pugRuntime: false`', function (done) {
-    var alias = require('rollup-plugin-alias')
+    const alias = require('rollup-plugin-alias')
 
     rollup({
-      entry: 'fixtures/runtime_import/main.js',
+      input: 'fixtures/runtime_import/main.js',
       plugins: [
         _pug({
-          pugRuntime: false
+          pugRuntime: false,
         }),
         alias({
-          pug_runtime: '../dist/runtime.es.js'
-        })
-      ]
+          pug_runtime: '../dist/runtime.es.js',
+        }),
+      ],
     })
-    .then(function (bundle) {
-      executeBundle(bundle, 'runtime_import')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'runtime_import'))
+      .then(done, done)
   })
 
   it('allows to skip runtime import with the `inlineRuntimeFunctions` option', function (done) {
     rollup({
-      entry: 'fixtures/custom_runtime/main.js',
+      input: 'fixtures/custom_runtime/main.js',
       plugins: [_pug({
         inlineRuntimeFunctions: true,
-        sourceMap: false
-      })]
+        sourceMap: false,
+      })],
     })
-    .then(function (bundle) {
-      debugger
-      executeBundle(bundle, 'inline_runtime')
-      done()
-    })
-    .catch(done)
+      .then((bundle) => executeBundle(bundle, 'inline_runtime'))
+      .then(done, done)
   })
 
 })