+++ typings/rollup-pluginutils/index.d.ts
@@ -1,0 +1,5 @@
+declare module "rollup-pluginutils" {
+  export const createFilter: (
+    include: string | string[] | undefined,
+    exclude: string | string[] | undefined) => (id: string) => boolean
+}